# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Allam Journal API",
    "version": "1.0.0",
    "description": "Single-journal OJS workflow. JSON commands and multipart uploads require a session-bound CSRF token. See README for login and resource policies. Response bodies are projections documented in architecture.md."
  },
  "servers": [
    {
      "url": "/",
      "description": "The backend serving this documentation"
    }
  ],
  "paths": {
    "/api/auth/csrf": {
      "get": {
        "summary": "GET /api/auth/csrf",
        "security": [],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid input"
          },
          "403": {
            "description": "Permission or CSRF denied"
          },
          "409": {
            "description": "Workflow, version, or uniqueness conflict"
          }
        },
        "tags": [
          "Authentication & identity"
        ],
        "operationId": "get_api_auth_csrf"
      }
    },
    "/api/auth/register": {
      "post": {
        "summary": "POST /api/auth/register",
        "security": [],
        "responses": {
          "400": {
            "description": "Invalid input"
          },
          "403": {
            "description": "Permission or CSRF denied"
          },
          "409": {
            "description": "Workflow, version, or uniqueness conflict"
          },
          "201": {
            "description": "Created resource with id"
          }
        },
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-TOKEN",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Fetch /api/auth/csrf using the same session cookie."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserInput"
              }
            }
          }
        },
        "tags": [
          "Authentication & identity"
        ],
        "operationId": "post_api_auth_register"
      }
    },
    "/api/me": {
      "get": {
        "summary": "GET /api/me",
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid input"
          },
          "403": {
            "description": "Permission or CSRF denied"
          },
          "409": {
            "description": "Workflow, version, or uniqueness conflict"
          }
        },
        "tags": [
          "Authentication & identity"
        ],
        "operationId": "get_api_me"
      }
    },
    "/api/me/orcid/authorize": {
      "post": {
        "summary": "POST /api/me/orcid/authorize",
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid input"
          },
          "403": {
            "description": "Permission or CSRF denied"
          },
          "409": {
            "description": "Workflow, version, or uniqueness conflict"
          }
        },
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-TOKEN",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Fetch /api/auth/csrf using the same session cookie."
          }
        ],
        "tags": [
          "Authentication & identity"
        ],
        "operationId": "post_api_me_orcid_authorize"
      }
    },
    "/api/me/orcid/callback": {
      "get": {
        "summary": "GET /api/me/orcid/callback",
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid input"
          },
          "403": {
            "description": "Permission or CSRF denied"
          },
          "409": {
            "description": "Workflow, version, or uniqueness conflict"
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "state",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "Authentication & identity"
        ],
        "operationId": "get_api_me_orcid_callback"
      }
    },
    "/api/submissions": {
      "post": {
        "summary": "POST /api/submissions",
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid input"
          },
          "403": {
            "description": "Permission or CSRF denied"
          },
          "409": {
            "description": "Workflow, version, or uniqueness conflict"
          },
          "201": {
            "description": "Created resource with id"
          }
        },
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-TOKEN",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Fetch /api/auth/csrf using the same session cookie."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftInput"
              }
            }
          }
        },
        "tags": [
          "Submissions & workflow"
        ],
        "operationId": "post_api_submissions"
      },
      "get": {
        "summary": "GET /api/submissions",
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid input"
          },
          "403": {
            "description": "Permission or CSRF denied"
          },
          "409": {
            "description": "Workflow, version, or uniqueness conflict"
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 100000,
              "default": 0
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 25
            }
          }
        ],
        "tags": [
          "Submissions & workflow"
        ],
        "operationId": "get_api_submissions"
      }
    },
    "/api/submissions/{id}": {
      "get": {
        "summary": "GET /api/submissions/{id}",
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid input"
          },
          "403": {
            "description": "Permission or CSRF denied"
          },
          "409": {
            "description": "Workflow, version, or uniqueness conflict"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "Submissions & workflow"
        ],
        "operationId": "get_api_submissions_id"
      }
    },
    "/api/submissions/{id}/metadata": {
      "put": {
        "summary": "PUT /api/submissions/{id}/metadata",
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid input"
          },
          "403": {
            "description": "Permission or CSRF denied"
          },
          "409": {
            "description": "Workflow, version, or uniqueness conflict"
          },
          "204": {
            "description": "Successful command"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-TOKEN",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Fetch /api/auth/csrf using the same session cookie."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MetadataInput"
              }
            }
          }
        },
        "tags": [
          "Authentication & identity"
        ],
        "operationId": "put_api_submissions_id_metadata"
      }
    },
    "/api/submissions/{id}/transitions": {
      "post": {
        "summary": "POST /api/submissions/{id}/transitions",
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid input"
          },
          "403": {
            "description": "Permission or CSRF denied"
          },
          "409": {
            "description": "Workflow, version, or uniqueness conflict"
          },
          "204": {
            "description": "Successful command"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-TOKEN",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Fetch /api/auth/csrf using the same session cookie."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionInput"
              }
            }
          }
        },
        "tags": [
          "Submissions & workflow"
        ],
        "operationId": "post_api_submissions_id_transitions"
      }
    },
    "/api/submissions/{id}/section-editor": {
      "put": {
        "summary": "PUT /api/submissions/{id}/section-editor",
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid input"
          },
          "403": {
            "description": "Permission or CSRF denied"
          },
          "409": {
            "description": "Workflow, version, or uniqueness conflict"
          },
          "204": {
            "description": "Successful command"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-TOKEN",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Fetch /api/auth/csrf using the same session cookie."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserAssignment"
              }
            }
          }
        },
        "tags": [
          "Submissions & workflow"
        ],
        "operationId": "put_api_submissions_id_section_editor"
      }
    },
    "/api/submissions/{id}/staff": {
      "post": {
        "summary": "POST /api/submissions/{id}/staff",
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid input"
          },
          "403": {
            "description": "Permission or CSRF denied"
          },
          "409": {
            "description": "Workflow, version, or uniqueness conflict"
          },
          "204": {
            "description": "Successful command"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-TOKEN",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Fetch /api/auth/csrf using the same session cookie."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffInput"
              }
            }
          }
        },
        "tags": [
          "Submissions & workflow"
        ],
        "operationId": "post_api_submissions_id_staff"
      }
    },
    "/api/submissions/{id}/review-mode": {
      "put": {
        "summary": "PUT /api/submissions/{id}/review-mode",
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid input"
          },
          "403": {
            "description": "Permission or CSRF denied"
          },
          "409": {
            "description": "Workflow, version, or uniqueness conflict"
          },
          "204": {
            "description": "Successful command"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-TOKEN",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Fetch /api/auth/csrf using the same session cookie."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModeInput"
              }
            }
          }
        },
        "tags": [
          "Submissions & workflow"
        ],
        "operationId": "put_api_submissions_id_review_mode"
      }
    },
...
    "/api/reviews": {
      "get": {
        "summary": "GET /api/reviews",
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid input"
          },
          "403": {
            "description": "Permission or CSRF denied"
          },
          "409": {
            "description": "Workflow, version, or uniqueness conflict"
          }
        },
        "tags": [
          "Peer review"
        ],
        "operationId": "get_api_reviews"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "JSESSIONID"
      }
    },
    "schemas": {
      "UserInput": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 16,
            "maxLength": 72
          },
          "name": {
            "type": "string",
            "maxLength": 200
          }
        },
        "required": [
          "email",
          "password",
          "name"
        ]
      },
...
      "DraftInput": {
        "type": "object",
        "properties": {
          "sectionId": {
            "type": "string"
          },
          "language": {
            "type": "string",
            "example": "en"
          },
          "checklist": {
            "type": "boolean"
          },
          "metadata": {
            "$ref": "#/components/schemas/Metadata"
          }
        },
        "required": [
          "sectionId",
          "language",
          "checklist",
          "metadata"
        ]
      },
...
      "UserAssignment": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          }
        },
        "required": [
          "userId"
        ]
      },
...
    }
  },
...
}